Add LanguageSettingInput
