+++ client/components/admin/settings/inputs/SelectSettingInput.js
@@ -35,6 +35,7 @@ export function SelectSettingInput({
 		</Field.Row>
 		<SelectInput
 			data-qa-setting-id={_id}
+			id={_id}
 			value={value}
 			placeholder={placeholder}
 			disabled={disabled}
